svg/painting/discrete-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'color-interpolation',
  from: 'initial',
  to: 'auto'
});

test_no_interpolation({
  property: 'fill-rule',
  from: 'initial',
  to: 'evenodd'
});

test_no_interpolation({
  property: 'marker-end',
  from: 'initial',
  to: 'url("#ref")'
});

test_no_interpolation({
  property: 'marker-mid',
  from: 'initial',
  to: 'url("#ref")'
});

test_no_interpolation({
  property: 'marker-start',
  from: 'initial',
  to: 'url("#ref")'
});
</script>
